AI/requirements.txt
numpy>=1.21.0
openai>=1.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0
